+++ Evergreen/Public/Get-VercelHyper.ps1
@@ -1,0 +1,44 @@
+Function Get-VercelHyper {
+    <#
+        .SYNOPSIS
+            Returns the latest Vercel Hyper version number and download.
+
+        .DESCRIPTION
+            Returns the latest Vercel Hyper version number and download.
+
+        .NOTES
+            Author: Aaron Parker
+            Twitter: @stealthpuppy
+        
+        .LINK
+            https://github.com/aaronparker/Evergreen
+
+        .EXAMPLE
+            Get-VercelHyper
+
+            Description:
+            Returns the latest Vercel Hyper version number and download for each platform.
+    #>
+    [Alias("Get-Atom")]
+    [OutputType([System.Management.Automation.PSObject])]
+    [CmdletBinding()]
+    Param()
+
+    # Get application resource strings from its manifest
+    $res = Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1]
+    Write-Verbose -Message $res.Name
+
+    # Pass the repo releases API URL and return a formatted object
+    $params = @{
+        Uri          = $res.Get.Uri
+        MatchVersion = $res.Get.MatchVersion
+        Filter       = $res.Get.MatchFileTypes
+    }
+    $object = Get-GitHubRepoRelease @params
+    If ($object) {
+        Write-Output -InputObject $object
+    }
+    Else {
+        Write-Warning -Message "$($MyInvocation.MyCommand): Failed to return a usable object from the repo."
+    }
+}