+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## VERSION
+
+* Adds `Get-AdobeBrackets`, `Get-Fork`, `Get-MicrosoftVisualStudio`, `Get-VercelHyper`
+* Updates manifest for `MicrosoftWvdRemoteDesktop` to ensure evergreen source URLs used for resolving downloads
+* Updates manifest for `MicrosoftVisualStudioCode`
+
 ## 2101.256
 
 * Adds `Get-Terminals`, `Get-PeaZipPeaZip`, `Get-Slack`, `Get-MicrosoftWindowsPackageManagerClient`, `Get-KeePassXCTeamKeePassXC`, `Get-SumatraPDFReader`